pipeline/requirements.txt
requests==2.32.3
feedparser==6.0.11
python-dotenv==1.0.1
beautifulsoup4==4.13.3
lxml==5.3.0
Pillow==11.0.0
supabase==2.9.1
